+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **doc:** add tags and shields
 - **add:** sinon
 - **fix:** refactor for better testing
+- **add:** tests for utils and index
 
 ### [v0.0.3] - 2016-11-22
 - **add:** rect support